revert the timeout as it really isn't indicative of what it was reporting break up some of the android test client stuff so that we are not explicitly coupled to the activity rule add some gradle vars for android plugin config and update to api26 target add/edit some readme files html report tab counts

+++ plugin/src/main/kotlin/com/trevjonez/kontrast/task/HtmlReportTask.kt
@@ -25,7 +25,6 @@ import org.gradle.api.DefaultTask
 import org.gradle.api.tasks.OutputDirectory
 import org.gradle.api.tasks.TaskAction
 import java.io.File
-import java.util.concurrent.TimeUnit
 
 open class HtmlReportTask : DefaultTask() {
 
@@ -38,31 +37,32 @@ open class HtmlReportTask : DefaultTask() {
 
     @TaskAction
     fun invoke() {
+        if(project.gradle.startParameter.taskNames.contains(name))
+            throw IllegalArgumentException("Html report task is ran automatically after the kontrast test task(s), do not invoke explicitly.")
+
         if (outputDir.exists())
             if (!outputDir.deleteRecursively()) throw IllegalStateException("Unable to delete old report directory")
 
         if (!outputDir.mkdirs()) throw IllegalStateException("Unable to create new report directory")
 
-        testCases.timeout(10, TimeUnit.MILLISECONDS)
-                .onErrorResumeNext { Single.error(IllegalArgumentException("Html report task is ran automatically after the kontrast test task(s), do not invoke explicitly.")) }
-                .map {
-                    it.map { testCase ->
-                        TestCaseReportInput(testCase.className, testCase.methodName,
-                                            testCase.testKey, testCase.inputExtras,
-                                            testCase.keyExtras, testCase.status,
-                                            File(outputDir, "images"))
-                                .also { reportCase ->
-                                    if (testCase.inputImage.exists())
-                                        FileUtils.copyFile(testCase.inputImage, reportCase.inputImage)
+        testCases.map {
+            it.map { testCase ->
+                TestCaseReportInput(testCase.className, testCase.methodName,
+                                    testCase.testKey, testCase.inputExtras,
+                                    testCase.keyExtras, testCase.status,
+                                    File(outputDir, "images"))
+                        .also { reportCase ->
+                            if (testCase.inputImage.exists())
+                                FileUtils.copyFile(testCase.inputImage, reportCase.inputImage)
 
-                                    if (testCase.keyImage.exists())
-                                        FileUtils.copyFile(testCase.keyImage, reportCase.keyImage)
+                            if (testCase.keyImage.exists())
+                                FileUtils.copyFile(testCase.keyImage, reportCase.keyImage)
 
-                                    if (testCase.diffImage.exists())
-                                        FileUtils.copyFile(testCase.diffImage, reportCase.diffImage)
-                                }
-                    }
-                }
+                            if (testCase.diffImage.exists())
+                                FileUtils.copyFile(testCase.diffImage, reportCase.diffImage)
+                        }
+            }
+        }
                 .map { ReportIndex(outputDir, variantName, it) }
                 .blockingGet()
                 .write()
